+++ src/modules/ToDoList.js
@@ -1,0 +1,9 @@
+class ToDoList {
+  constructor(desc, completed = false, index) {
+    this.desc = desc;
+    this.completed = completed;
+    this.index = index;
+  }
+}
+
+export default ToDoList;
